+++ NEWS.md
@@ -2,7 +2,7 @@
 
 - Added a function that checks Databricks variables for ODBC connections: `check_databricks_odbc()`.
 - Updated `wd_pcon_lad_la_rgn_ctry` by adding three digit local authority codes (also known as old_la_codes) to the data.
-- Added old_la_codes data which is a lookup table of local authorities with their names, nine digit codes and old three digit codes.
+- Added old_la_codes which is an internal lookup table of local authorities with their names, nine digit codes and 'old' three digit codes. It is used to add 'old' 3 digit la codes to  `wd_pcon_lad_la_rgn_ctry`. 
 - Updated `fetch_las()` so that old_la_codes are included in the output.
 
 # dfeR 1.0.1